Moved functions to better place
