+++ sentry_sdk/integrations/redis/modules/caches.py
@@ -1,16 +1,50 @@
 """
-Code used for the Queries modules in Sentry
+Code used for the Caches module in Sentry
 """
 
 from sentry_sdk._types import TYPE_CHECKING
-from sentry_sdk.consts import SPANDATA
+from sentry_sdk.consts import OP, SPANDATA
 from sentry_sdk.integrations.redis.utils import _get_safe_key
 
 
 if TYPE_CHECKING:
     from sentry_sdk.integrations.redis import RedisIntegration
     from sentry_sdk.tracing import Span
-    from typing import Any
+    from typing import Any, Optional
+
+
+def _get_op(name):
+    # type: (str) -> Optional[str]
+    op = None
+    if name.lower().startswith("get"):
+        op = OP.CACHE_GET_ITEM
+    elif name.lower().startswith("set"):
+        op = OP.CACHE_SET_ITEM
+
+    return op
+
+
+def _compile_cache_span_properties(integration, redis_command, args, kwargs):
+    # type: (RedisIntegration, str, tuple[Any], dict[str, Any]) -> dict[str, Any]
+    key = _get_safe_key(redis_command, args, kwargs)
+
+    is_cache_key = False
+    for prefix in integration.cache_prefixes:
+        if key.startswith(prefix):
+            is_cache_key = True
+            break
+
+    properties = {
+        "op": _get_op(redis_command),
+        "description": _get_cache_span_description(
+            integration, redis_command, args, kwargs
+        ),
+        "key": key,
+        "redis_command": redis_command,
+        "is_cache_key": is_cache_key,
+    }
+
+    return properties
 
 
 def _get_cache_span_description(integration, command_name, args, kwargs):
