+++ sentry_sdk/integrations/redis/modules/queries.py
@@ -1,10 +1,10 @@
 """
-Code used for the Queries modules in Sentry
+Code used for the Queries module in Sentry
 """
 
 from sentry_sdk._types import TYPE_CHECKING
-from sentry_sdk.consts import SPANDATA
-from sentry_sdk.integrations.redis.utils import _get_safe_command
+from sentry_sdk.consts import OP, SPANDATA
+from sentry_sdk.integrations.redis.utils import _get_safe_command, _get_safe_key
 from sentry_sdk.utils import capture_internal_exceptions
 
 
@@ -13,6 +13,21 @@ if TYPE_CHECKING:
     from sentry_sdk.integrations.redis import RedisIntegration
     from sentry_sdk.tracing import Span
     from typing import Any
+
+
+def _compile_db_span_properties(integration, redis_command, args, kwargs):
+    # type: (RedisIntegration, str, tuple[Any], dict[str, Any]) -> dict[str, Any]
+    key = _get_safe_key(redis_command, args, kwargs)
+    description = _get_db_span_description(integration, redis_command, *args)
+
+    properties = {
+        "op": OP.DB_REDIS,
+        "description": description,
+        "redis_command": redis_command,
+        "key": key,
+    }
+
+    return properties
 
 
 def _get_db_span_description(integration, command_name, *args):
